A bunch of updates, part from Manuel Novoa III (such as more long long support), and other updates by me (better cross platform, cross-compiler, etc, support.  Now compiles with 2.0.x kernels for armnommu.

+++ libc/stdlib/malloc/alloc.c
@@ -12,7 +12,7 @@ void *
 calloc_dbg(size_t num, size_t size, char * function, char * file, int line)
 {
 	void * ptr;
-	fprintf(stderr, "calloc of %d bytes at %s @%s:%d = ", num*size, function, file, line);
+	fprintf(stderr, "calloc of %ld bytes at %s @%s:%d = ", (long)(num*size), function, file, line);
 	ptr = calloc(num,size);
 	fprintf(stderr, "%p\n", ptr);
 	return ptr;
@@ -26,7 +26,7 @@ void *
 malloc_dbg(size_t len, char * function, char * file, int line)
 {
 	void * result;
-	fprintf(stderr, "malloc of %d bytes at %s @%s:%d = ", len, function, file, line);
+	fprintf(stderr, "malloc of %ld bytes at %s @%s:%d = ", (long)len, function, file, line);
 	result = malloc(len);
 	fprintf(stderr, "%p\n", result);    
 	return result;
